Metadata guideline for tagging Titanic original maps
it would be nice to have some distinction for maps intentionally created within the Titanic community. right now, web does not have an adequate distinction between original Titanic maps and exclusives (from other servers) because there's no way to mark it. this can help with that I think. 

maybe wording could be improved to be more specific about when one should apply Titanic tag? it's up to mapper choice though so it's intended to be a guideline